+++ R/zzz.R
@@ -1,6 +1,6 @@
 #' @importFrom Rdpack reprompt
 
 .onAttach <- function(libname, pkgname) {
-  packageStartupMessage("Smoothed Empirical Likelihood estimation version 0.0.12 (2024-06-13).")
+  packageStartupMessage("Smoothed Empirical Likelihood estimation version 0.0.13 (2025-02-05).")
   packageStartupMessage("This is *NOT* a stable version. Core functions subject to change.")
 }